add BOARD_GKI_NONAB_COMPAT for non-A/B compatability with GKI
The directory structure of several ramdisks are different depending on
whether the target is A/B or non-A/B. When the target needs to be
compatible with the GKI however, which is always built as A/B, it is
necessary to build these ramdisks using the A/B file hierarchy. Provide
a flag to enable this override when necessary.

Bug: 161579556
Change-Id: I1cf9a0ef5a338b0b5152113533b62804f3088c83