Add ANGLE.apk to the base system
This CL re-enables building ANGLE as parta of the base system. This was
done up until the recent merge of goog/master to aosp/master, and is
required as part of enabling ANGLE as the default OpenGL ES driver for
Cuttlefish.

Bug: b/168625635
Test: atest CtsAngleIntegrationHostTestCases
Change-Id: I12c8db290ed7dc6b213b6f1c22692666dcdf143f
Merged-In: Ide4e3d8f78d3748d557a47ecd153d66694fd1e3a