*-ndk.so libraries are added to the VNDK set.
We will replace the ndk_platform backend with the ndk backend. The
former hasn't been deleted yet because there already are references to
it. When the existing references are all replaced with the ndk backend,
we will remove the ndk_platform backend and the corresponding libraries
will be removed from this list.

Bug: 161456198
Test: m

Change-Id: Ia32d82d6e32c367bdf1f92151c7da7408e8bcbfa