Don't require current.txt order.
There have been a lot of merge conflicts because of this file (and a
few cases of people needing to merge changes internally because of
it). Though, we really don't care about the order of the file. So,
compare ignoring order. If there is a merge conflict in the future,
we can change the order slightly, and we can gradually clean up the
order over time, as needed (really, it's not important).

Bug: N/A
Test: file reordered as build-time check
Change-Id: I955a2bd78872ff6d975bc494f10b2ed5e532c680