Retry: Enable mainline prebuilts on tm-dev branch.
Except for santizer and framework Java coverage tests as they require
specialized prebuilts built from a matching configuration and we do not
provide them.

And products that use dex2oat(d)s as they are not available as a
prebuilt.

And products that cross compile to arm64 and linux_bionic on the host
as prebuilts for that are not available.

And atv/tv products as they break when building against prebuilts.

Bug: 230478310
Bug: 223592962
Test: presubmit
Merged-In: I3ed091156c4b39dad6eaf578144ac3d070025988
Merged-In: I9731400acfade95b3eab8f2d99fddb4cf856aa91
Merged-In: I43fa4ae8db51012eaf78d70fe65e9e0a3f7eca16
Change-Id: Ib01ba9e2e3c2225c88d52b6811edf04aff02c81e