Allow building super.img on targets without BOARD_BUILD_SUPER_IMAGE_BY_DEFAULT set.
With this change, if BOARD_BUILD_SUPER_IMAGE_BY_DEFAULT is unset, a
target will be declared for super.img so it becomes possible to build it
(e.g. with `m superimage`), but it will not be part of the default `droid`
target. This is useful for building a `super.img` on targets that do
not normally need one without needing to go through a slow `m dist` step.

Change-Id: I599e93b3fc41fc2c0642761fcda757c40e2736d6