Tune R8 shrinking defaults for system_server
Disable R8 shrinking by default if there are additional
SYSTEM_SERVER_JARS that may be downstream from services.jar, or if
the jar ordering is unreliable.

The longer term solution will be to feed R8 the SYSTEM_SERVER_JARS
explicitly during the shrinking phase, but for now this avoids
potential issues with such dowstream jars. Note that specific
products or targets can still opt-in explicitly even if there may be
ordering issues in SYSTEM_SERVER_JARS.

Bug: 240588226,235484969
Test: m + lunch various Arc/Wear/Auto targets
Change-Id: I1d824e1e61de65604123b65b25b192a8a32fd0c1