BOARD_VNDK_VERSION always by default
There is a lot of pending cleanup pending on always having
the VNDK enabled. For instance, when some devices always
build with the core variant, it is impossible to have code
excluded from the system variant, since vendor code on
non-VNDK devices wouldn't be able to build.

This CL merely disables the flag. When it lands, there is a
lot of cleanup that would follow.

Bug: 122954981
Test: ¯\_(ツ)_/¯ removing untested paths
Change-Id: Ia9380fe54729ce1ab37cf0f5e8512be45cd550c7