temporarily disable roboleaf_tests.sh
When I enabled the envsetup_tests target on CI, this script was failing.
Rather than break the build again by enabling CI on a potentially
failing script, let's just disable this for now, enable the CI target,
and then this script can be enabled and tested in presubmits before
submitting.

Bug: 244771601
Test: build/make/tests/roboleaf_tests.sh
Change-Id: I7c3469b49258e37c0ff9b49dd8805031216a7570