Update naming of platforms subdir
The build system previously named the platforms subdir by
PLATFORM_VERSION, which is 13 for android 13, aka T, so the subdir
name would be android-13.

Android studio installs the SDK into a subdir identified by the api
level, so android-33 for android T, and also the extension api level
for extension SDKs, e.g. android-33-ext4.

Update the build system subdir name to match the android studio install
path, in order to avoid confusion and make it simpler to just unzip
a built SDK into the platforms dir, avoiding the need to provide a
package.xml file for studio to install it correctly.

Bug: 254023524
Test: m sdk sdk_repo dist, verify subdir is android-33 in:
   $ zipinfo -1 out/dist/android-sdk_eng.hansson_linux-x86.zip \
     | cut -f 1-3 -d '/' | grep platforms | sort | uniq
Change-Id: I1d3a7fd4c480859c28cdcbcb8c0cf2800ac4a1d7