Adds the gpu_counter_producer command.
This is a small binary that will load a GPU vendor provided shared
library called libgpudataproducer.so, look for the "start" symbol in
that library, and call it. This can be used to start the GPU hardware
counter Perfetto producer that will allow tracing of GPU profiling data
via Perfetto.

Test: manual
Bug: b/273907878
Ignore-AOSP-First: change in topic with internal-first project
Change-Id: I450e159c4d4c323ade90ffa46759239a7bcd4aac