Make sort as part of the inner command
It turns out, in Cog workspaces, the order of files returned by "find"
command is inherently non-deterministic (they use an absl::HashMap which
intentionally tries to NOT guarantee sort order). This results in
varying order for the inner variable, which actually causes invalidation
and regeneration of the build graph (I'm guessing when "y" changes in
$(sort $(y)), it causes an invalidation in Make). Hence made the sort a
part of the inner command itself.

Tested: Ran the build with Cog and ensured graph wasn't regenerated
across reruns.
Bug: b/276397558

Change-Id: Ie772572048785793067f74c08ac3994ef6cbaa43