Split ANGLE supported and enabled into different Makefiles.
ANGLE binaries should be added to the PRODUCT_PACKAGES instead. This
patch also splits the angle.mk to angle_supported.mk and
angle_default.mk so that when a device just wants to ship ANGLE binaries
to coexist with native GLES drivers, it only needs to inherit the
angle_supported.mk and when a device wants to ship ANGLE as the default
system GLES drivers, it only needs to inherit the angle_default.mk. When
ANGLE is the default system GLES driver, make sure persist.graphics.egl
is defined so that ANGLE is loaded properly.

Bug: 270994705
Test: atest CtsAngleIntegrationHostTestCases
Change-Id: I3de4a40b62b39c91c8f6df56d68755830b303f10
Merged-In: I3de4a40b62b39c91c8f6df56d68755830b303f10