Add adevice_fingerprint to eng builds.
We are working on a replacement for "adb sync" and it requires something
running on the device to send back a fingerprint of the filesystem.
This binary is intendend to be part of developers going forward so we
are adding it in all eng/userdebug builds.
The binary is 830k bytes.

Test: m install-clean ; m ; acloud create --local-image ; adb shell ls -l /system/bin/adevice_fingerprint

Change-Id: I98359cc37cffeffc7ffbfd61d6bc9a51ba318a99