Fixes for python 3.11
Bazel previously had a bug where the source directory of the entrypoint
python script would be added to the path. They added PYTHONSAFEPATH
to the enviornment to fix it, which was introduced in python 3.11.

Now that python is being updated to 3.11, the fix is in, which breaks
some scripts that don't properly specify import folders.

Add the import folders where required.

Bug: 278602456
Test: ./build/bazel/ci/bp2build.sh
Change-Id: I5b87f83247dfdfcb8bdb4f962466cbf522cc29c6