Move vendor module into base_vendor
There is a vendor module android.hidl.memory@1.0-impl.vendor which is
defined in the base_system. Based on the definition of base_system.mk
and base_vendor.mk, this module is better to be located in the
base_vendor.mk. This change moves the module from base_system into
base_vendor.

Bug: 301344301
Test: Cuttlefish build succeeded
Change-Id: I6dba5ec2380028b1b2c629f901db996c518e3f72