Configure bootclasspath for sdk product
The SDK product is currently used to build the SDK sources, which
are the full sources of the bootclasspath. For this to work, the
bootclasspath needs to be configured correctly, so inherit that config.

In the future, it would probably make sense for the sources to be built
together with a device -- the emulator, at which point this config can
be removed.

Bug: 151360309
Test: lunch sdk; m sdk sdk_repo dist
Change-Id: I66b09cd42249c4d5b76f06478781a90108e006e4