Reapply "release_config: build flags can be lists"
Allow a build flag definition to indicate that its value should be the
concatentation of assignements, rather than the final assigned value. In
this case, the "default" value from the flag definition is always
present as the start of the list.

The initial use case for this is RELEASE_ACONFIG_VALUE_SETS, where we
need apply multiple definition files that should be processed to arrive
at the final value.

This reverts commit b05eaac092155d918260f7dd703e3b4fefa95027.

Bug: b/302593603, b/304814040
Test: manual

Change-Id: I7370c509ceb3952f7feb2351673d8f2ba86d704b