Allow duplicate release config maps and build flag declarations
Only include a release_config_map.mk file the first time it appears in
the list.

If the flag declaration is identical to one already declared, do not
raise an error.

This eases the process of migrating where flags are declared.

Bug: b/328781415
Bug: b/328495189
Test: manual, TH
Ignore-AOSP-First: larger topic on git_main, will cp to AOSP.
Change-Id: I18316e769438f62071723771762d6e89be98b813