Drop HOST_CROSS_OS special-case in source vs prebuilt mainline selection
We currently do not have many products that cross compile to
linux_bionic, and none of them use google apexes.

One exception is `mainline_sdk` product, but this product exclusively
builds using the `mainline` release config. This release config contains
the necessary RELEASE_APEX_CONTRIBUTIONS_* to provide sources during the
build, so we do not need this additional special-case here

Bug: 308188056
Test: lunch trout_arm64-next-userdebug && get_build_var
PRODUCT_PACKAGES; verified that google apexes do not appear there before
this CL, effectively making this CL a no-op for that product

Change-Id: Ibfdc742b6ae01242be22a3d3f6bb3c0cf551705f