Allow AVF remote attestation to be disabled
This CL adds a variable in makefile that allows the
AVF remote attestation feature to be disabled easily.

The remote attestation feature is opt-in by default, as it
is a strongly recommended feature from Android V. It will
be disabled if PRODUCT_AVF_REMOTE_ATTESTATION_DISABLED is
set to true in the Makefile, otherwise it will be enabled
when the feature flag
RELEASE_AVF_ENABLE_REMOTE_ATTESTATION is on.

Bug: 331269709
Bug: 341292601
Test: run VM remote attestation on Pixel 6 and check it is
disabled

Change-Id: Id48ebd0eacf0d7a6c69b9100aee5ab6e4f112fe5