Create PRODUCT_APEX_BOOT_JARS_PREBUILT_INCLUDED_LIST and add framework-pdf in the list.
From M-2024-08, the framework-pdf will be added as part of MediaProvider
apex and has it usage covered to S+ platform. Thus, this framework-pdf
jar will also be included in prebuilt BCP from S. For older platforms to
be buildable with M-2024-08 prebuilt, we would need to add framework-pdf
to it Apex boot jars.

Bug: 343766613
Test: build aosp_arm64-userdebug in udc-dev
Test: build cf_x86_phone-userdebug with V prebuilts in udc-dev
Change-Id: I19698e5f0d180d7848db67432a13f414f4b7ccb5