Migrate ramdisk_node_list configuration to Android.bp
Removed the hardcoded assignment of `RAMDISK_NODE_LIST` in
`build/core/Makefile`. Instead, a filegroup definition in Android.bp is
now used to manage `ramdisk_node_list` and its export to the make
variable. This change enhances flexibility and simplifies the build
process by eliminating the need for a fixed installation path.

Bug: 359423289
Test: m -j $(PRODUCT_OUT)/ramdisk.img
Change-Id: I5aea108d426a615fa951c368c16f7cdea73252b1