Add framework-connectivity-b to the boot classpath
This is a preparation CL for moving VCN from platform to Tethering
module. As the first step, VCN will be built into it's own
java_sdk_library on the platform and be loaded into the boot
classpath. Followup CLs will move VCN to a mainline module and
thus add framework-connectivity-b to PRODUCT_APEX_BOOT_JARS with a flag.

This CL also improves the dup check of PRODUCT_APEX_BOOT_JARS and
PRODUCT_BOOT_JARS. Previously, it checkes whether the apex jar name
is a substring of PRODUCT_BOOT_JARS. It will trigger false alert
when there is a "framework-connectivity" in PRODUCT_APEX_BOOT_JARS
and a "framework-connectivity-b" in PRODUCT_BOOT_JARS

Bug: 369703242
Test: build and boot
Flag: EXEMPT create new empty libraries; no functional change
Change-Id: I9e5d77c8f5b7b014bde84fd9cc1997852196af37