Return null if key is not found in map
The major change of this commit is return null if the key if not found
in the flag maps. The key cannot be found because of the following
reasons

1. the package is not in the container
1. the flag is not in the container

These are valid query for a map to test whether a key exists in a map.

Test: atest aconfig_storage_read_api.test.java aconfig_storage_file.test.java
Bug: 368135182
Flag: NONE files for flag
Change-Id: Ibcf05c6113e5780db06e52e0dcd8ed3bcdd3384c