core/Makefile: Make PACK_DESKTOP_{RECOVERY,UPDATE}_IMAGE into product variable
Make the PACK_DESKTOP_UPDATE_IMAGE and PACK_DESKTOP_RECOVERY_IMAGE
variables into product variables PRODUCT_BUILD_DESKTOP_UPDATE_IMAGE and
PRODUCT_BUILD_DESKTOP_RECOVERY_IMAGE so that board specific mk files can
override the value of these variables. The original variables can only
be appended to or overridden with 'override' because we use
'inherit-product'. The use of 'inherit-product' causes the test in this
Makefile to always succeed because the variables are set to 'true' by
default. The only other way to reset these variables seems to be to use
'override <var> = ' in the board's product mk file, which isn't
preferred vs. using a product variable.

Bug: 368171802
Test: m pack-update-image; \
      m pack-recovery-image
Change-Id: Iba3bc8bf1715785662260ad7dfc0927cb3e52477