Support ADB product keys purging with a dedicated flag
Instead of checking the build variant look for the value of a dedicated
flag to decide if we need to do the purging. The value of
RELEASE_BUILD_PURGE_PRODUCT_ADB_KEYS is set to true on user and false on
userdebug/eng across all releases. Because of that this patch does not
bring any functional changes.

BUG: b/373580227
Test: Add $(warning foo) to the conditional block. Verify that the output
      of lunch husky-${aosp_current,trunk_staging}-{user,userdebug,eng}
      is as expected.

Change-Id: Ice6042025e9d2767e3eee158960eaab46f9b7fd4