Update vendor_api_levels.go with the vintf finalization
vendor_api_levels.go provides corresponding SDK version for vendor
API levels. After the finalization, we will have the next vendor API
level for the trunk_staging build.
We may change the map file when the version is changed later, but
must provide an SDK version for the next vendor API level for the
trunk_staging build.

Ignore-AOSP-First: finalization script is an exception
Bug: 331696132
Test: build/make/tools/finalization/build-step-0.sh
Change-Id: I6788add9e2177cc9d2fc63f54e032f5fc3ecf9b9