Use RELEASE_PLATFORM_SDK_VERSION_FULL instead of RELEASE_PLATFORM_SDK_MINOR_VERSION
RELEASE_PLATFORM_SDK_MINOR_VERSION has been replaced by
RELEASE_PLATFORM_SDK_VERSION_FULL (because we don't want to refer to the
minor version in isolation; *SDK_VERSION_FULL encodes both major and
minor version in the same variable).

If the value of RELEASE_PLATFORM_SDK_VERSION_FULL is the empty string
(the default), the build will synthesize a value by concatenating
RELEASE_PLATFORM_SDK_VERSION and ".0", resulting in e.g. "35.0".

If the value of RELEASE_PLATFORM_SDK_VERSION_FULL, it is assumed to be
on the format "<major>.<minor>", and the <major> part must match the
value of RELEASE_PLATFORM_SDK_VERSION.

Bug: 380235511
Test: atest CtsOsTestCases:BuildTest
Ignore-AOSP-First: minor versions do not exist in AOSP yet
Change-Id: Ibec28ed9317b5d561250335a26ef760a860f34f9