Create aconfig_flags for all containers.
Some targets do not create images for some non-system partitions, but
include them in the system partition. Even in such cases, create the
aconfig_flags in the etc directory of the original container.
For example, system_ext aconfig_flags.pb can be installed in
system/system_ext/etc/aconfig_flags.pb.

With this change, add system_ext to the aconfig paths along with the
other partitions.

Bug: 385213580
Test: lunch gsi_x86_64_soong_system-trunk_staging-userdebug && m
Change-Id: I660eb2e3d55850680ca610b8b756495d1769de4c