Update test mode codegen for nextfood
There is a cc test mode codegen lib that failed to build when enabling
read from new storage. The underlying issue is that in nextfood, all
flags are RO, even the fallback should just return a hardcoded value.
Before this change, it still tries to read from storage file which
caused a build failure.

Test: m
Bug: b/370046610
Ignore-AOSP-First: need to go with a git_main only change, will CP it back to aosp afterwards.
Change-Id: I4267ac9eb1fc6b068a8808e76d698a12936e31b6