record-finalized-flags: introduce new tool
Add a new tool, record-finalized-flags. The tool will be used to create
a list of all @FlaggedApi flags that participate in the next API
finalization so we can add tooling to prevent reusing "spent" flags in
future @FlaggedApi annotations.

This CL only adds the new soong module to allow soong to call the new
tool: the actual implementation will land in future CLs.

Bug: 377676163
Test: m all_aconfig_declarations
Change-Id: I81b3dd0158764c9936fba63a29acdbe80d1d44b0