Remove output_file before generating it
aosp/3478832 recently switched some fsv_meta files to symlinks.
The symlink creation and file writing operations in most programming
languages don't work properly when the destination is already exists
and is of the opposite type. (symlinking on top of an existing file/
symlink would fail, file writing would write through the symlink)
Remove the output file first.

This wasn't noticed origionally because it only happens on certain
types of incremental builds, and is also fixed if you run an
`m installclean`.

Bug: 394404628
Test: m
Change-Id: I7f900d52b81bb696e597f20f20a092987e42db2e