Deprecate TARGET_BUILD_APPS=all
We'd like to convert more of the packaging logic to soong as we work
to get rid of make. TARGET_BUILD_APPS=all makes that tricky, because
we wouldn't know if there was a make-defined app that needed to be
included in the apps package. It doesn't appear to be used, so
deprecate it.

Bug: 376727180
Test: Presubmits
Change-Id: Ie701cbd847260959a537938e2c29c696d107b9b0