Declare inputs of $PRODUCT_OUT/system_server.zip
 https://r.android.com/3062109 switched system_server.zip generation to
read the files from an intermediates directory assembled by a soong
singleton, and not from $PRODUCT_OUT. However, it did not update the
inputs. This silently relied on the fact that depending on the system
server jars in PRODUCT_OUT would assemble the jar in the soong
intermediates directory

This CL makes the input explicit.

Bug: 396162261
Test: m $PRODUCT_OUT/system_server.zip
Test: aninja -t query $PRODUCT_OUT/system_server.zip
(verified that out/soong/system_server_dexjars/services.jar is listed in
inputs)

Change-Id: I087597e26562250c9870cf636cd457ad3a1394c8