Remove requirement that fsv_meta file is next to its input file
In soong-built partitions, we will generate fsv_meta files before
a partition staging directory is generated. This means that the source
file will be some module-specific intermediates path, and the dest
file will be a path local to the filesystem module. Relax the
requirement that fsv_meta files are next to the files they protect.

Bug: 394404628
Test: m aosp_shared_system_image, verified it didn't change before/after
Change-Id: I6bbbd23aa93b15220c5b2302cb0b229e8f8de69c